test: Station 관련 Rest 기능 분리
